pyproject.toml
[project]
name = "fraud-detector"
version = "0.1.0"
description = "E2E Predictive MLOps Demo — Fraud Detection on Google Cloud"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-storage>=2.18.0",
    "google-cloud-aiplatform>=1.72.0",
    "pandas>=2.2.0",
    "pyarrow>=17.0.0",
    "db-dtypes>=1.3.0",
    "scipy>=1.17.0",
    "xgboost>=2.1.0",
    "scikit-learn>=1.5.0",
    "joblib>=1.4.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
pipelines = [
    "kfp>=2.9.0",
    "google-cloud-pipeline-components>=2.17.0",
]
notebook = [
    "jupyterlab>=4.0.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "C", "B", "UP", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
